+++ src/Helpers/ImprovementHelper.php
@@ -28,7 +28,7 @@ class ImprovementHelper
             'keep' => '+%s%% max population',
             'forges' => '+%s%% offensive power',
             'walls' => '+%s%% defensive power',
-            'spires' => '+%s%% offensive wizard power & mana production<br>+%s%% protection from spells',
+            'spires' => '+%s%% offensive wizard power & mana production<br>+%s%% spell damage reduction',
             'harbor' => '+%s%% food production<br>+%s%% boat production & protection',
         ];
 
@@ -44,7 +44,7 @@ class ImprovementHelper
             'keep' => "{$improvementName}: population bonus maxes out at +30% base",
             'forges' => "{$improvementName}: offense bonus maxes out at +30% base",
             'walls' => "{$improvementName}: defense bonus maxes out at +30% base",
-            'spires' => "{$improvementName}: wizard power and mana bonuses max out at +60% base<br><br>Protection: spell protection bonus is increased by 50%, maxes out at +50%, and cannot be modified by masonries",
+            'spires' => "{$improvementName}: wizard power and mana bonuses max out at +60% base<br><br>Protection: spell damage reduction bonus is increased by 50%, maxes out at +50%, and cannot be modified by masonries",
             'harbor' => "{$improvementName}: food bonus maxes out at +60% base<br><br>Protection: boat bonuses are increased by 50%, max out at +50%, and cannot be modified by masonries",
         ];